requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.3.1
coloredlogs==15.0.1
crewai==1.6.1
crewai-tools==1.6.1
cryptography==46.0.3
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
et_xmlfile==2.0.0
fastuuid==0.14.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.12.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.2.1
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.13.0
Jinja2==3.1.6
jiter==0.11.1
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
lance-namespace==0.2.1
lance-namespace-urllib3-client==0.2.1
lancedb==0.25.3
langchain-core==1.1.1
langchain-google-genai==3.2.0
langsmith==0.4.56
litellm==1.80.8
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.23.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
nodeenv==1.9.1
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.9.0
openpyxl==3.1.5
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-exporter-otlp-proto-http==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
platformdirs==4.5.1
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.5.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.6
pypdfium2==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytube==15.0.0
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
tqdm==4.67.1
ty==0.0.1a32
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.3.0
uuid_utils==0.12.0
uv==0.9.16
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
youtube-transcript-api==1.2.3
zipp==3.23.0
zstandard==0.25.0
<task_progress>
- [x] Create requirements file
- [x] Add all installed packages with compatible versions
- [ ] Save requirements.txt
</task_progress>
</write_to_file>
